# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tool Rent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List all tools
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, rented, maintenance]
        - name: shopId
          in: query
          schema:
            type: number
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
    post:
      operationId: createTool
      tags: [tools]
      summary: Create a tool
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolRequest"
      responses:
        "201":
          description: Tool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, refunded]
        - name: rentalId
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Payment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /analytics/dashboard:
    get:
      operationId: getDashboardAnalytics
      tags: [analytics]
      summary: Get dashboard analytics
      security:
        - bearerAuth: []
      parameters:
        - name: shopId
          in: query
          schema:
            type: number
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
      responses:
        "200":
          description: Dashboard analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardAnalytics"
  /analytics/tools:
    get:
      operationId: getToolAnalytics
      tags: [analytics]
      summary: Get tool usage analytics
      security:
        - bearerAuth: []
      parameters:
        - name: shopId
          in: query
          schema:
            type: number
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
      responses:
        "200":
          description: Tool analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolAnalytics"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, shop_owner, worker]
      required:
        - name
        - phone
        - password
        - role
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
        - phone
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, shop_owner, worker, customer]
        shopId:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - role
        - isActive
        - createdAt
...
    Tool:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        category:
          type: string
        shopId:
          type: number
        shopName:
          type: string
        pricePerDay:
          type: number
        depositAmount:
          type: number
        status:
          type: string
          enum: [available, rented, maintenance]
        qrCode:
          type: string
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - shopId
        - pricePerDay
        - depositAmount
        - status
        - qrCode
        - createdAt
    CreateToolRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        shopId:
          type: number
        pricePerDay:
          type: number
        depositAmount:
          type: number
        imageUrl:
          type: string
      required:
        - name
        - category
        - shopId
        - pricePerDay
        - depositAmount
...
    ToolListResponse:
      type: object
      properties:
        tools:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - tools
        - total
        - page
        - limit
...
    Rental:
      type: object
      properties:
        id:
          type: number
        toolId:
          type: number
        toolName:
          type: string
        customerId:
          type: number
        customerName:
          type: string
        shopId:
          type: number
        workerId:
          type: number
        status:
          type: string
          enum: [active, returned, overdue]
        rentalPrice:
          type: number
        depositAmount:
          type: number
        totalAmount:
          type: number
        startedAt:
          type: string
          format: date-time
        returnedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        damageNote:
          type: string
        damageCost:
          type: number
      required:
        - id
        - toolId
        - customerId
        - shopId
        - status
        - rentalPrice
        - depositAmount
        - totalAmount
        - startedAt
...
    Payment:
      type: object
      properties:
        id:
          type: number
        rentalId:
          type: number
        amount:
          type: number
        method:
          type: string
          enum: [click, payme, paynet, cash]
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        type:
          type: string
          enum: [rental, deposit, deposit_refund, damage_deduction]
        transactionId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - rentalId
        - amount
        - method
        - status
        - type
        - createdAt
    CreatePaymentRequest:
      type: object
      properties:
        rentalId:
          type: number
        amount:
          type: number
        method:
          type: string
          enum: [click, payme, paynet, cash]
        type:
          type: string
          enum: [rental, deposit, deposit_refund, damage_deduction]
      required:
        - rentalId
        - amount
        - method
        - type
    PaymentListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - payments
        - total
        - page
        - limit
    DashboardAnalytics:
      type: object
      properties:
        totalRevenue:
          type: number
        activeRentals:
          type: number
        overdueRentals:
          type: number
        totalTools:
          type: number
        availableTools:
          type: number
        rentedTools:
          type: number
        totalCustomers:
          type: number
        revenueByPeriod:
          type: array
          items:
            $ref: "#/components/schemas/RevenuePoint"
        recentRentals:
          type: array
          items:
            $ref: "#/components/schemas/Rental"
      required:
        - totalRevenue
        - activeRentals
        - overdueRentals
        - totalTools
        - availableTools
        - rentedTools
        - totalCustomers
        - revenueByPeriod
        - recentRentals
    RevenuePoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        rentals:
          type: number
      required:
        - date
        - revenue
        - rentals
    ToolAnalytics:
      type: object
      properties:
        topRentedTools:
          type: array
          items:
            $ref: "#/components/schemas/ToolUsageStat"
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryStat"
      required:
        - topRentedTools
        - categoryBreakdown
    ToolUsageStat:
      type: object
      properties:
        toolId:
          type: number
        toolName:
          type: string
        totalRentals:
          type: number
        totalRevenue:
          type: number
      required:
        - toolId
        - toolName
        - totalRentals
        - totalRevenue
    CategoryStat:
      type: object
      properties:
        category:
          type: string
        totalTools:
          type: number
        totalRentals:
          type: number
      required:
        - category
        - totalTools
        - totalRentals