# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tool Rent Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  phone: string;
  email?: string;
  role: UserRole;
  shopId?: number;
  isActive: boolean;
  createdAt: Date;
}